Function minimisation

The main functionality of ConsensusBasedX.jl is function minimisation via Consensus-Based Optimisation. It assumes you have defined a function f(x::AbstractVector) that takes a single vector argumemt x of length D = length(x).

For instance, if D = 2, you can minimise f by running:

minimise(f, D = 2)

Full-code example.

Note

You must always provide D.

Using a config object

For more advanced usage, you will select several options. You can pass these as extra keyword arguments to minimise, or you can create a NamedTuple called config and pass that:

config = (; D = 2)
minimise(f, config)

Full-code example.

Note

If you pass a Dict instead, it will be converted to a NamedTuple automatically.

Aliases

ConsensusBasedX.jl also defines minimize, optimise, and optimize. These are all aliases of minimise.

Maximisation

ConsensusBasedX.jl also defines maximise (and its alias, maximize) for convenience. If you call

maximise(f, D = 2)

or

config = (; D = 2)
maximise(f, config)

maximise will attempt to define g(x) = -f(x) and call minimise(g, config).

Full-code examples are provided for the keyword and config approaches.

Method reference

ConsensusBasedX.maximise

ConsensusBasedX.minimise

ConsensusBasedX.maximise — Function

maximise(f; keywords...)

maximise(config::NamedTuple, f)

Maximise the function f using Consensus-Based Optimisation.

Attempts to define x -> -f(x) and calls the minimise routine. This might be better handled directly by the user (see Maximisation).

maximize is an alias for maximise.

See also minimise.

ConsensusBasedX.minimise — Function

minimise(f; keywords...)

minimise(f, config::NamedTuple)

Minimise the function f using Consensus-Based Optimisation (see Function minimisation).

You must specify the dimension D of the problem. Other paramters (e.g. the number of particles N or the number of ensembles M can also be specified; see Summary of options.

minimize, optimise, or optimize are aliases for minimise.

Examples

minimise(f, D = 2)

config = (; D = 2);
minimise(f, config)

minimise(f, D = 2, N = 20)

config = (; D = 2, N = 20);
minimise(f, config)