Low-level interface

Internally, the minimise routine relies on two constructs: the ParticleDynamic and a ConsensusBasedXMethod.

ConsensusBasedXLowLevel

ConsensusBasedX.jl defines the ConsensusBasedXLowLevel submodule for the convenience of advanced users. It can be imported by

using ConsensusBasedX, ConsensusBasedX.ConsensusBasedXLowLevel

ParticleDynamic

The ParticleDynamic struct defines the evolution in time of particles, and is agnostic of the specific method in question. Its functionality currently serves ConsensusBasedOptimisation, but can be extended to other methods.

ConsensusBasedX.ParticleDynamic — Type

ParticleDynamic

Fields:

method<:ConsensusBasedXMethod, the optimisation method.

Δt::Float64, the time step.

ParticleDynamicCache

ParticleDynamic requires a cache, ParticleDynamicCache. This can be constructed with:

ConsensusBasedX.construct_particle_dynamic_cache — Function

construct_particle_dynamic_cache(
  config::NamedTuple,
  X₀::AbstractArray,
  particle_dynamic::ParticleDynamic,
)

A constructor helper for ParticleDynamicCache. Calls ConsensusBasedX.construct_method_cache to construct the corresponding ConsensusBasedXMethodCache.

The full reference is:

ConsensusBasedX.ParticleDynamicCache — Type

ParticleDynamicCache

It is strongly recommended that you do not construct ParticleDynamicCache by hand. Instead, use ConsensusBasedX.construct_particle_dynamic_cache.

Fields:

mode should be set to ParticleMode.

parallelisation<:Parallelisations, the parallelisation mode.

method_cache<:ConsensusBasedXMethodCache, a cache for the method field of ParticleDynamic.

D::Int, the dimension of the problem.

N::Int, the number of particles per ensemble.

M::Int, the number of ensembles.

X, the particle array.

dX, the time derivative array.

Δt::Float64, the time step.

rootΔt::Float64, the square root of the time step.

root2Δt::Float64, the square root of twice the time step.

max_iterations::Float64, the maximum number of iterations.

max_time::Float64, the maximal time.

iteration::Vector{Int}, the vector containing the iteration count per ensemble.

ConsensusBasedOptimisation

The ConsensusBasedOptimisation struct (of type ConsensusBasedXMethod) defines the details of the consensus-based optimisation method (function evaluations, consensus point...).

ConsensusBasedX.ConsensusBasedOptimisation — Type

ConsensusBasedOptimisation

Fields:

f, the objective function.

correction<:ConsensusBasedXCorrection, a correction term.

α::Float64, the the exponential weight parameter.

λ::Float64, the drift strengh.

σ::Float64, the noise strengh.

ConsensusBasedOptimisationCache

ConsensusBasedOptimisation requires a cache, ConsensusBasedOptimisationCache (of type ConsensusBasedXMethodCache). This can be constructed with:

ConsensusBasedX.construct_method_cache — Function

construct_method_cache(
  config::NamedTuple,
  X₀::AbstractArray,
  method::ConsensusBasedXMethod,
  particle_dynamic::ParticleDynamic,
)

A constructor helper for ConsensusBasedXMethodCache.

Note that this method is called automatically by ConsensusBasedX.construct_particle_dynamic_cache.

The full reference is:

ConsensusBasedX.ConsensusBasedOptimisationCache — Type

ConsensusBasedOptimisationCache{T}

It is strongly recommended that you do not construct ConsensusBasedOptimisationCache by hand. Instead, use ConsensusBasedX.construct_method_cache.

Fields:

consensus::Vector{Vector{T}}, the consensus point of each ensemble.

consensus_energy::Vector{T}, the energy (value of the objective function) of each consensus point.

consensus_energy_previous::Vector{T}, the previous energy.

distance::Vector{Vector{T}}, the distance of each particle to the consensus point.

energy::Vector{Vector{T}}, the energy of each particle.

exponents::Vector{Vector{T}}, an exponent used to compute logsums.

logsums::Vector{T}, a normalisation factor for weights.

weights::Vector{Vector{T}}, the exponential weight of each particle.

energy_threshold::Float64, the energy threshold.

energy_tolerance::Float64, the energy tolerance.

max_evaluations::Float64, the maximum number of f evaluations.

evaluations::Vector{Int}, the current number of f evaluations.